+++ app/src/commonMain/kotlin/com/crosspaste/db/paste/PasteDao.kt
@@ -3,61 +3,39 @@ package com.crosspaste.db.paste
 import app.cash.sqldelight.Query
 import app.cash.sqldelight.coroutines.asFlow
 import com.crosspaste.Database
-import com.crosspaste.app.AppFileType
 import com.crosspaste.app.AppInfo
-import com.crosspaste.config.CommonConfigManager
-import com.crosspaste.paste.CurrentPaste
-import com.crosspaste.paste.PasteCollection
 import com.crosspaste.paste.PasteData
 import com.crosspaste.paste.PasteExportParam
 import com.crosspaste.paste.PasteState
-import com.crosspaste.paste.PasteTag
 import com.crosspaste.paste.SearchContentService
-import com.crosspaste.paste.item.PasteFiles
 import com.crosspaste.paste.item.PasteItem
-import com.crosspaste.paste.item.PasteItemProperties
-import com.crosspaste.paste.plugin.process.PasteProcessPlugin
 import com.crosspaste.path.UserDataPathProvider
-import com.crosspaste.task.TaskBuilder
 import com.crosspaste.task.TaskSubmitter
 import com.crosspaste.utils.DateUtils
 import com.crosspaste.utils.LoggerExtension.logExecutionTime
-import com.crosspaste.utils.getFileUtils
 import com.crosspaste.utils.ioDispatcher
 import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.map
-import kotlinx.serialization.json.booleanOrNull
-import kotlinx.serialization.json.jsonPrimitive
 import kotlinx.coroutines.withContext
-import kotlin.time.ExperimentalTime
 
 class PasteDao(
     private val appInfo: AppInfo,
-    private val commonConfigManager: CommonConfigManager,
-    private val currentPaste: CurrentPaste,
     private val database: Database,
-    private val pasteProcessPlugins: List<PasteProcessPlugin>,
     private val searchContentService: SearchContentService,
     private val taskSubmitter: TaskSubmitter,
     private val userDataPathProvider: UserDataPathProvider,
 ) {
 
-    companion object {
-        private val fileUtils = getFileUtils()
-    }
-
     private val logger = KotlinLogging.logger {}
 
     private val pasteDatabaseQueries = database.pasteDatabaseQueries
 
-    private val tagDatabaseQueries = database.tagDatabaseQueries
-
     private val markDeleteBatchNum = 50L
 
-    private fun getNoDeletePasteDataBlock(id: Long): PasteData? {
+    fun getNoDeletePasteDataBlock(id: Long): PasteData? {
         return pasteDatabaseQueries.getPasteData(
             id,
             listOf(PasteState.LOADING.toLong(), PasteState.LOADED.toLong()),
@@ -226,86 +204,13 @@ class PasteDao(
             .flowOn(ioDispatcher)
     }
 
-    fun getAllTagsFlow(): Flow<List<PasteTag>> {
-        return tagDatabaseQueries.getAllTags(PasteTag::mapper)
-            .asFlow()
-            .map { it.executeAsList() }
-            .catch { e ->
-                logger.error(e) { "Error executing getAllTagsFlow query: ${e.message}" }
-                emit(listOf())
-            }
-            .flowOn(ioDispatcher)
-    }
-
-    suspend fun getMaxSortOrder(): Long = withContext(ioDispatcher) {
-        tagDatabaseQueries.maxSortOrder().executeAsOne()
-    }
-
-    suspend fun createPasteTag(name: String, color: Long): Long = withContext(ioDispatcher) {
-        database.transactionWithResult {
-            val maxSortOrder = tagDatabaseQueries.maxSortOrder().executeAsOne()
-            val newSortOrder = maxSortOrder + 1
-            tagDatabaseQueries.createTag(name, color, newSortOrder)
-            tagDatabaseQueries.getLastId().executeAsOne()
-        }
-    }
-
-    suspend fun updatePasteTagName(id: Long, name: String) = withContext(ioDispatcher) {
-        tagDatabaseQueries.updateTagName(name, id)
-    }
-
-    suspend fun updatePasteTagColor(id: Long, color: Long) = withContext(ioDispatcher) {
-        tagDatabaseQueries.updateTagColor(color, id)
-    }
-
-    private fun pinPasteTag(pasteDataId: Long, pasteTagId: Long) {
-        tagDatabaseQueries.pinPasteTag(pasteDataId, pasteTagId)
-    }
-
-    private fun unPinPasteTag(pasteDataId: Long, pasteTagId: Long) {
-        tagDatabaseQueries.unPinPasteTag(pasteDataId, pasteTagId)
-    }
-
-    fun switchPinPasteTagBlock(pasteDataId: Long, pasteTagId: Long) {
-        database.transaction {
-            val pinned = tagDatabaseQueries.isPinnedPasteTag(pasteDataId, pasteTagId).executeAsOne()
-            if (pinned) {
-                unPinPasteTag(pasteDataId, pasteTagId)
-            } else {
-                pinPasteTag(pasteDataId, pasteTagId)
-            }
-        }
-    }
-
-    fun getPasteTagsBlock(pasteDataId: Long): List<Long>  {
-        return tagDatabaseQueries.getPasteTags(pasteDataId).executeAsList()
-    }
-
-    fun deletePasteTagBlock(id: Long) {
-        tagDatabaseQueries.deleteTag(id)
-    }
-
-    @OptIn(ExperimentalTime::class)
-    private fun TaskBuilder.markDeleteSameHash(
-        newPasteDataId: Long,
-        newPasteDataType: Int,
-        newPasteDataHash: String,
-    ) {
-        if (newPasteDataHash.isEmpty()) {
-            return
-        }
-
-        val idList = pasteDatabaseQueries.getSameHashPasteDataIds(
-            newPasteDataHash,
-            newPasteDataType.toLong(),
+    fun getSameHashPasteDataIds(hash: String, pasteType: Int, excludeId: Long): List<Long> {
+        return pasteDatabaseQueries.getSameHashPasteDataIds(
+            hash,
+            pasteType.toLong(),
             DateUtils.getOffsetDay(days = -1),
-            newPasteDataId,
+            excludeId,
         ).executeAsList()
-
-        database.transaction {
-            pasteDatabaseQueries.markDeletePasteData(idList)
-            addDeletePasteTasks(idList)
-        }
     }
 
     suspend fun markDeleteByCleanTime(
@@ -477,161 +382,6 @@ class PasteDao(
     suspend fun getDistinctSources(): List<String> = withContext(ioDispatcher) {
         pasteDatabaseQueries.getDistinctSources(appInfo.appInstanceId)
             .executeAsList()
-    }
-
-    suspend fun releaseRemotePasteData(
-        pasteData: PasteData,
-        tryWritePasteboard: (PasteData) -> Unit,
-    ): Result<Unit> {
-        return runCatching {
-            val remotePasteDataId = pasteData.id
-            val isFileType = pasteData.isFileType()
-            val existIconFile: Boolean? =
-                pasteData.source?.let {
-                    fileUtils.existFile(userDataPathProvider.resolve("$it.png", AppFileType.ICON))
-                }
-
-            val pasteState = if (isFileType) {
-                PasteState.LOADING
-            } else {
-                PasteState.LOADED
-            }
-
-            val id = createPasteData(pasteData, pasteState)
-
-            taskSubmitter.submit {
-                if (!isFileType) {
-                    markDeleteSameHash(id, pasteData.pasteType, pasteData.hash)
-                    addRenderingTask(id, pasteData.getType())
-                    tryWritePasteboard(pasteData)
-                } else {
-                    val pasteFiles = pasteData.getPasteItem(PasteFiles::class)
-
-                    if (pasteFiles == null) {
-                        logger.warn { "File-type paste $id has no PasteFiles item, skipping" }
-                        return@submit
-                    }
-
-                    val fileSize = pasteFiles.size
-                    val maxBackupFileSize =
-                        fileUtils.bytesSize(
-                            commonConfigManager.getCurrentConfig().maxBackupFileSize,
-                        )
-
-                    val syncToDownload =
-                        fileSize > maxBackupFileSize ||
-                            pasteData.pasteAppearItem?.extraInfo
-                                ?.get(PasteItemProperties.SYNC_TO_DOWNLOAD)
-                                ?.jsonPrimitive?.booleanOrNull == true
-
-                    val pasteCoordinate = pasteData.getPasteCoordinate(id)
-                    val pasteAppearItem = pasteData.pasteAppearItem
-                    val pasteCollection = pasteData.pasteCollection
-
-                    val newPasteAppearItem = pasteAppearItem?.bind(pasteCoordinate, syncToDownload)
-                    val newPasteCollection = pasteCollection.bind(pasteCoordinate, syncToDownload)
-
-                    val newPasteData = pasteData.copy(
-                        id = id,
-                        pasteAppearItem = newPasteAppearItem,
-                        pasteCollection = newPasteCollection,
-                    )
-
-                    updateFilePath(newPasteData)
-
-                    addPullFileTask(id, remotePasteDataId)
-                    addRelaySyncTask(id, newPasteData.appInstanceId)
-                }
-
-                existIconFile?.let {
-                    addPullIconTask(id, it)
-                }
-            }
-        }.onFailure { e ->
-            logger.error(e) { "Release remote paste data failed" }
-        }
-    }
-
-    suspend fun releaseRemotePasteDataWithFile(
-        id: Long,
-        tryWritePasteboard: (PasteData) -> Unit,
-    ): Result<Unit> = withContext(ioDispatcher) {
-        runCatching {
-            taskSubmitter.submit {
-                database.transactionWithResult {
-                    pasteDatabaseQueries.updatePasteDataState(PasteState.LOADED.toLong(), id)
-                    getNoDeletePasteDataBlock(id)
-                }?.let {
-                    markDeleteSameHash(id, it.pasteType, it.hash)
-                    addRelaySyncTask(id, it.appInstanceId)
-                    tryWritePasteboard(it)
-                }
-            }
-        }.onFailure { e ->
-            logger.error(e) { "Release remote paste data with file failed" }
-        }
-    }
-
-    suspend fun releaseLocalPasteData(id: Long, pasteItems: List<PasteItem>) = withContext(ioDispatcher) {
-        getLoadingPasteData(id)?.let { pasteData ->
-            var pasteAppearItems = pasteItems
-            for (pastePlugin in pasteProcessPlugins) {
-                pasteAppearItems = pastePlugin.process(
-                    pasteData.getPasteCoordinate(),
-                    pasteAppearItems,
-                    pasteData.source,
-                )
-            }
-
-            if (pasteAppearItems.isEmpty()) {
-                markDeletePasteData(id)
-                return@let
-            }
-
-            val size = pasteAppearItems.sumOf { it.size }
-            val maxFileSize = pasteAppearItems.filter { it is PasteFiles }
-                .maxByOrNull { it.size }
-                ?.size ?: 0
-            // first item as pasteAppearItem
-            // remaining items as pasteContent
-            val firstItem: PasteItem = pasteAppearItems.first()
-            val remainingItems: List<PasteItem> = pasteAppearItems.drop(1)
-
-            val hash = firstItem.hash
-            val pasteType = firstItem.getPasteType()
-
-            val change = database.transactionWithResult {
-                pasteDatabaseQueries.updatePasteDataToLoaded(
-                    pasteAppearItem = firstItem.toJson(),
-                    pasteCollection = PasteCollection(remainingItems).toJson(),
-                    pasteType = pasteType.type.toLong(),
-                    pasteSearchContent = searchContentService.createSearchContent(
-                        pasteData.source,
-                        firstItem.getSearchContent(),
-                    ),
-                    size = size,
-                    hash = hash,
-                    id = id,
-                )
-                pasteDatabaseQueries.change().executeAsOne() > 0
-            }
-
-            if (change) {
-                currentPaste.setPasteId(id)
-                taskSubmitter.submit {
-                    addRenderingTask(id, pasteType)
-                    addSyncTask(id, pasteData.appInstanceId, maxFileSize)
-
-                    if (pasteType.isFile() || pasteType.isImage()) {
-                        if ((firstItem as PasteFiles).isRefFiles()) {
-                            markDeleteSameHash(id, pasteType.type, hash)
-                        }
-                    } else {
-                        markDeleteSameHash(id, pasteType.type, hash)
-                    }
-                }
-            }
-        }
     }
 
     suspend fun getPasteResourceInfo(favorite: Boolean? = null): PasteResourceInfo = withContext(ioDispatcher) {
