+++ app/src/desktopMain/kotlin/com/crosspaste/paste/WindowsPasteboardService.kt
@@ -3,7 +3,6 @@ package com.crosspaste.paste
 import com.crosspaste.app.DesktopAppWindowManager
 import com.crosspaste.config.CommonConfigManager
 import com.crosspaste.config.DesktopConfigManager
-import com.crosspaste.db.paste.PasteDao
 import com.crosspaste.notification.NotificationManager
 import com.crosspaste.paste.item.PasteItem
 import com.crosspaste.paste.item.PasteText
@@ -38,7 +37,7 @@ class WindowsPasteboardService(
     override val notificationManager: NotificationManager,
     override val pasteConsumer: TransferableConsumer,
     override val pasteProducer: TransferableProducer,
-    override val pasteDao: PasteDao,
+    override val pasteReleaseService: PasteReleaseService,
     private val platform: Platform,
     override val soundService: SoundService,
     override val sourceExclusionService: DesktopSourceExclusionService,
